Giant reorganization and refactor cleanup
