+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 
 members = [
-    "app-libs/exchange-oracle",
+    "app-libs/oracle",
     "app-libs/sgx-runtime",
     "app-libs/stf",
     "cli",
