--- app-libs/exchange-oracle/src/traits.rs
@@ -1,4 +1,0 @@
-pub trait TradingInfo {
-	type TradingPair;
-	type ExchangeRate;
-}
