+++ app-libs/oracle/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ita-exchange-oracle"
+name = "ita-oracle"
 version = "0.9.0"
 authors = ["Integritee AG <hello@integritee.network>"]
 edition = "2021"
