+++ app-libs/oracle/src/certificates/open_meteo_root.pem
