+++ app-libs/oracle/src/error.rs
