+++ app-libs/oracle/src/types.rs
@@ -58,3 +58,4 @@ impl TradingPair {
 /// TODO Fix https://github.com/integritee-network/pallets/issues/71 and get it from https://github.com/integritee-network/pallets.git
 /// Teeracle types
 pub type ExchangeRate = U32F32;
+// pub type Coordinate = U32F32;
